requirements.txt
fastapi==0.109.0
g4f[all]==0.2.7.1
psycopg2-binary==2.9.9
uvicorn==0.25.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.1